backend/requirements.txt
Django>=3.2,<4
djangorestframework>=3.12
djangorestframework-simplejwt>=5.2
django-cors-headers>=4.0
argon2-cffi>=23.1
